Rename and reorganize files
