+++ lib/geoflutterfire2.dart
@@ -1,5 +1,5 @@
 library geoflutterfire2;
 
-export 'src/collection.dart';
+export 'src/collection/collection.dart';
 export 'src/geoflutterfire2.dart';
-export 'src/point.dart';
+export 'src/models/point.dart';
