+++ lib/src/utils/arrays.dart
